FirebaseService go deeper
